Moves QiskitOrgMenu component to components/layouts folder
