+++ layouts/default.vue
@@ -12,7 +12,7 @@
 <script lang="ts">
 import Vue from 'vue'
 import { Component } from 'vue-property-decorator'
-import QiskitOrgMenu from '~/components/menus/QiskitOrgMenu.vue'
+import QiskitOrgMenu from '~/components/layouts/QiskitOrgMenu.vue'
 import PageFooter from '~/components/footers/PageFooter.vue'
 
 @Component({
